requirements.txt
alembic==1.13.1
asgiref==3.8.1
black==24.4.2
bleach==6.1.0
blinker==1.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==42.0.7
cssbeautifier==1.15.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==6.0.0
dj3-cloudinary-storage==0.0.6
Django==5.0.7
django-allauth==0.57.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-filter==24.2
django-mailjet==0.3.1
django-phonenumber-field==8.0.0
django-summernote==0.8.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djlint==1.34.1
EditorConfig==0.12.4
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
geographiclib==2.0
geopy==2.4.1
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsbeautifier==1.15.1
json5==0.9.25
mailjet-rest==1.3.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.40
Pillow==8.2.0
platformdirs==4.2.2
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
sendgrid==6.11.0
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==1.26.18
webencodings==0.5.1
Werkzeug==3.0.2
whitenoise==5.3.0
